refactor: 초기 데이터 설정 파일 이동
