+++ src/test/java/nextstep/common/utils/DataLoader.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.common.utils;
 
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
@@ -18,5 +18,10 @@ public class DataLoader {
     public void loadData() {
         memberRepository.save(new Member("admin@email.com", "password", 20, List.of(RoleType.ROLE_ADMIN.name())));
         memberRepository.save(new Member("member@email.com", "password", 20, List.of(RoleType.ROLE_MEMBER.name())));
+
+        memberRepository.save(new Member("email1@email.com", "password", 20, List.of(RoleType.ROLE_MEMBER.name())));
+        memberRepository.save(new Member("email2@email.com", "password", 20, List.of(RoleType.ROLE_MEMBER.name())));
+        memberRepository.save(new Member("email3@email.com", "password", 20, List.of(RoleType.ROLE_MEMBER.name())));
+        memberRepository.save(new Member("email4@email.com", "password", 20, List.of(RoleType.ROLE_MEMBER.name())));
     }
 }
